+++ clients/vscode-hlasmplugin/src/hlasmListingServices.ts
@@ -17,11 +17,6 @@ import { EXTENSION_ID, initialBlanks, languageIdHlasmListing } from './constants
 
 const ordchar = /[A-Za-z0-9$#@_]/;
 
-// Symbol   Length   Value     Id    R Type Asm  Program   Defn References                     
-// A             4 00000000 00000001     A  A                 1   44    45    46    47    48    49    50    51    52    53 
-//                                                                54    55    56    57    58    59    60    61    62    63 
-//                                                                64    65    66    67    68    69    70    71    72    73 
-
 const listingStart = /^(.?)                                         High Level Assembler Option Summary                   .............   Page    1/;
 
 type RegexSet = {
@@ -33,28 +28,61 @@ type RegexSet = {
     ordinaryRefFirstLine: RegExp,
     ordinaryRefAltSecondLine: RegExp,
     ordinaryRefRest: RegExp,
+
+    externalRefFirstLine: RegExp,
+    externalRefSecondLine: RegExp,
+
+    dsectRefFirstLine: RegExp,
+    dsectRefSecondLine: RegExp,
+
+    usingMapLine: RegExp,
+
+    sectionAddrShort: RegExp,
+    sectionAddrLong: RegExp,
 };
 
 const withoutPrefix = {
-    objShortCode: /^.{33}( *\d+)\D/,
-    objLongCode: /^.{41}( *\d+)\D/,
+    objShortCode: /^(?:([0-9ABCDEF]{6})| {6}) (.{26})( *\d+)\D(?:([a-zA-Z$#@_][a-zA-Z$#@0-9_]*) )?/,
+    objLongCode: /^(?:([0-9ABCDEF]{8})| {8}) (.{32})( *\d+)\D(?:([a-zA-Z$#@_][a-zA-Z$#@0-9_]*) )?/,
     lineText: /^(?:(Return Code )|\*\* (ASMA\d\d\d[NIWES] .+)|((?:  |[CDR]-)Loc  Object Code    Addr1 Addr2  Stmt |(?:  |[CDR]-)Loc    Object Code      Addr1    Addr2    Stmt )|(.{111})Page +\d+)/,
     pageBoundary: /^.+(?:(High Level Assembler Option Summary)|(External Symbol Dictionary)|(Relocation Dictionary)|(Ordinary Symbol and Literal Cross Reference)|(Macro and Copy Code Source Summary)|(Dsect Cross Reference)|(Using Map)|(General Purpose Register Cross Reference)|(Diagnostic Cross Reference and Assembler Summary))/,
 
-    ordinaryRefFirstLine: /^(?:([a-zA-Z$#@_][a-zA-Z$#@0-9_]{0,7}) +(\d+) ([0-9A-F]{8}) [0-9A-F]{8} . .... ...  ....... +(\d+) +(\d.+|)|([a-zA-Z$#@_][a-zA-Z$#@0-9_]{8,}))/,
-    ordinaryRefAltSecondLine: /^( {9,})(\d+) ([0-9A-F]{8}) [0-9A-F]{8} . .... ...  ....... +(\d+) +(\d.+|)/,
+    ordinaryRefFirstLine: /^(?:([a-zA-Z$#@_][a-zA-Z$#@0-9_]{0,7}) +(\d+) ([0-9A-F]{8}) ([0-9A-F]{8}) (.) ..(.). ...  ....... +(\d+) +(\d.+|)|([a-zA-Z$#@_][a-zA-Z$#@0-9_]{8,}))/,
+    ordinaryRefAltSecondLine: /^( {9,})(\d+) ([0-9A-F]{8}) ([0-9A-F]{8}) (.) ..(.). ...  ....... +(\d+) +(\d.+|)/,
     ordinaryRefRest: /^ {60,}(\d.+)/,
+
+    externalRefFirstLine: /^([a-zA-Z$#@_][a-zA-Z$#@0-9_]{0,7}) +([A-Z]+) +([0-9A-F]+) ( {8}|[0-9A-F]{8}) ( {8}|[0-9A-F]{8})  ( {8}|[0-9A-F]{8}) |^([a-zA-Z$#@_][a-zA-Z$#@0-9_]{8,} +)/,
+    externalRefSecondLine: /^( +)([A-Z]+) +([0-9A-F]+) ( {9}|[0-9A-F]{8} )( {9}|[0-9A-F]{8} ) ( {8}|[0-9A-F]{8})/,
+
+    dsectRefFirstLine: /^(?:([a-zA-Z$#@_][a-zA-Z$#@0-9_]{0,7}) +([0-9A-F]{8}) +([0-9A-F]{8}) +(\d+)|([a-zA-Z$#@_][a-zA-Z$#@0-9_]{8,}))/,
+    dsectRefSecondLine: /^( +)([0-9A-F]{8}) +([0-9A-F]{8}) +(\d+)/,
+
+    usingMapLine: /^ *(\d+)  ([0-9A-F]{8})  ([0-9A-F]{8}) (?:USING|DROP|PUSH|POP) +/,
+
+    sectionAddrShort: /^ {15}[0-9A-F]{6} [0-9A-F]{6}$/,
+    sectionAddrLong: /^ {15}[0-9A-F]{8} [0-9A-F]{8}$/,
 };
 
 const withPrefix = {
-    objShortCode: /^..{33}( *\d+)\D/,
-    objLongCode: /^..{41}( *\d+)\D/,
+    objShortCode: /^.(?:([0-9ABCDEF]{6})| {6}) (.{26})( *\d+)\D(?:([a-zA-Z$#@_][a-zA-Z$#@0-9_]*) )?/,
+    objLongCode: /^.(?:([0-9ABCDEF]{8})| {8}) (.{32})( *\d+)\D(?:([a-zA-Z$#@_][a-zA-Z$#@0-9_]*) )?/,
     lineText: /^.(?:(Return Code )|\*\* (ASMA\d\d\d[NIWES] .+)|((?:  |[CDR]-)Loc  Object Code    Addr1 Addr2  Stmt |(?:  |[CDR]-)Loc    Object Code      Addr1    Addr2    Stmt )|(.{111})Page +\d+)/,
     pageBoundary: /^.+(?:(High Level Assembler Option Summary)|(External Symbol Dictionary)|(Relocation Dictionary)|(Ordinary Symbol and Literal Cross Reference)|(Macro and Copy Code Source Summary)|(Dsect Cross Reference)|(Using Map)|(General Purpose Register Cross Reference)|(Diagnostic Cross Reference and Assembler Summary))/,
 
-    ordinaryRefFirstLine: /^.(?:([a-zA-Z$#@_][a-zA-Z$#@0-9_]{0,7}) +(\d+) ([0-9A-F]{8}) [0-9A-F]{8} . .... ...  ....... +(\d+) +(\d.+|)|([a-zA-Z$#@_][a-zA-Z$#@0-9_]{8,}))/,
-    ordinaryRefAltSecondLine: /^.( {9,})(\d+) ([0-9A-F]{8}) [0-9A-F]{8} . .... ...  ....... +(\d+) +(\d.+|)/,
+    ordinaryRefFirstLine: /^.(?:([a-zA-Z$#@_][a-zA-Z$#@0-9_]{0,7}) +(\d+) ([0-9A-F]{8}) ([0-9A-F]{8}) (.) ..(.). ...  ....... +(\d+) +(\d.+|)|([a-zA-Z$#@_][a-zA-Z$#@0-9_]{8,}))/,
+    ordinaryRefAltSecondLine: /^.( {9,})(\d+) ([0-9A-F]{8}) ([0-9A-F]{8}) (.) ..(.). ...  ....... +(\d+) +(\d.+|)/,
     ordinaryRefRest: /^. {60,}(\d.+)/,
+
+    externalRefFirstLine: /^.([a-zA-Z$#@_][a-zA-Z$#@0-9_]{0,7}) +([A-Z]+) +([0-9A-F]+) ( {8}|[0-9A-F]{8}) ( {8}|[0-9A-F]{8})  ( {8}|[0-9A-F]{8}) |^([a-zA-Z$#@_][a-zA-Z$#@0-9_]{8,} +)/,
+    externalRefSecondLine: /^.( +)([A-Z]+) +([0-9A-F]+) ( {9}|[0-9A-F]{8} )( {9}|[0-9A-F]{8} ) ( {8}|[0-9A-F]{8})/,
+
+    dsectRefFirstLine: /^.(?:([a-zA-Z$#@_][a-zA-Z$#@0-9_]{0,7}) +([0-9A-F]{8}) +([0-9A-F]{8}) +(\d+)|([a-zA-Z$#@_][a-zA-Z$#@0-9_]{8,}))/,
+    dsectRefSecondLine: /^.( +)([0-9A-F]{8}) +([0-9A-F]{8}) +(\d+)/,
+
+    usingMapLine: /^. *(\d+)  ([0-9A-F]{8})  ([0-9A-F]{8}) (?:USING|DROP|PUSH|POP) +/,
+
+    sectionAddrShort: /^ {15}[0-9A-F]{6} [0-9A-F]{6}$/,
+    sectionAddrLong: /^ {15}[0-9A-F]{8} [0-9A-F]{8}$/,
 };
 
 const enum BoudnaryType {
@@ -73,7 +101,7 @@ const enum BoudnaryType {
     OtherBoundary,
 };
 
-function testLine(s: string, r: RegexSet): { type: BoudnaryType, capture: string } | undefined {
+function identifyBoundary(s: string, r: RegexSet): { type: BoudnaryType, capture: string } | undefined {
     const l = r.lineText.exec(s);
     if (!l) return undefined;
 
@@ -105,15 +133,31 @@ type Section = {
     end: number,
 };
 
+type ListingLine = {
+    listingLine: number,
+    address: number | undefined,
+};
+
+type CodeSection = Section & {
+    title: string,
+    codeStart: number,
+    dsect: boolean,
+    firstStmtNo: number | undefined,
+};
+
 type Listing = {
     start: number,
     end: number,
     hasPrefix: boolean,
     type?: 'long' | 'short';
     diagnostics: vscode.Diagnostic[],
-    statementLines: Map<number, number>,
+    statementLines: Map<number, ListingLine>,
     symbols: Map<string, Symbol>,
-    codeSections: (Section & { title: string, codeStart: number })[],
+    sections: Map<string, string>,
+    csects: { name: string, address: number, length: number }[],
+    codeSections: CodeSection[],
+    maxStmtNum: number,
+    sectionMap: (boolean | undefined)[],
 
     options?: Section,
     externals?: Section,
@@ -139,17 +183,32 @@ function updateSymbols(symbols: Map<string, Symbol>, symbol: Symbol) {
     else {
         s.defined = [...new Set([...s.defined, ...symbol.defined])];
         s.references = [...new Set([...s.references, ...symbol.references])];
+        s.referencesPure = [...new Set([...s.referencesPure, ...symbol.referencesPure])];
     }
 }
 
-class Symbol {
-    name: string = '';
-    defined: number[] = [];
-    references: number[] = [];
+type SymbolDetails = {
+    value: number,
+    sectionId: number,
+    reloc: boolean,
+    undefined: boolean,
+    loctr: boolean,
+}
 
-    computeReferences(includeDefinition: boolean) {
-        return includeDefinition ? [...new Set([...this.references, ...this.defined])] : this.references;
-    }
+type Symbol = {
+    name: string,
+    defined: number[],
+    references: number[],
+    referencesPure: number[],
+    details: SymbolDetails | undefined,
+}
+
+function computeReferences(symbol: Symbol, includeDefinition: boolean) {
+    return includeDefinition ? [...new Set([...symbol.references, ...symbol.defined])] : symbol.references;
+}
+
+function isCsectLike(s: string) {
+    return s == "SD" || s == "ED" || s == "PC" || s == "CM";
 }
 
 function processListing(doc: vscode.TextDocument, start: number, hasPrefix: boolean): { nexti: number, result: Listing } {
@@ -159,9 +218,13 @@ function processListing(doc: vscode.TextDocument, start: number, hasPrefix: bool
         end: start,
         hasPrefix,
         diagnostics: [],
-        statementLines: new Map<number, number>(),
+        statementLines: new Map<number, ListingLine>(),
         symbols: new Map<string, Symbol>(),
+        sections: new Map<string, string>(),
+        csects: [],
         codeSections: [],
+        maxStmtNum: 0,
+        sectionMap: [],
     };
     type ListingSections = Exclude<{ [key in keyof Listing]: Listing[key] extends (Section | undefined) ? key : never }[keyof Listing], undefined>;
     const updateCommonSection = <Name extends ListingSections>(name: Name, lineno: number): Listing[Name] => {
@@ -172,45 +235,54 @@ function processListing(doc: vscode.TextDocument, start: number, hasPrefix: bool
     };
     const enum States {
         Options,
+        ExternalRefs,
         Code,
         OrdinaryRefs,
+        DsectRefs,
+        UsingRef,
         Refs,
     };
 
     let symbol: Symbol | undefined;
-    let lastSection: { end: number } | undefined = undefined;
+    let lastListingSection: { end: number } | CodeSection | undefined = undefined;
     let lastTitle = '';
     let lastTitleLine = start;
+    let lastCsect: string | undefined = undefined;
+    let lastDsect: string | undefined = undefined;
+
+    const symbolCandidates: { name: string, stmtNumber: number }[] = [];
+    const knownCsectStmtNumbers: number[] = [];
+    const addrTransionsStmtNumbers: number[] = [];
 
     let state = States.Options;
     let i = start;
     main: for (; i < doc.lineCount; ++i) {
         const line = doc.lineAt(i);
-        const l = testLine(line.text, r);
-        if (l) {
-            if (lastSection) {
-                lastSection.end = i;
-                lastSection = undefined;
+        const b = identifyBoundary(line.text, r);
+        if (b) {
+            if (lastListingSection) {
+                lastListingSection.end = i;
+                lastListingSection = undefined;
             }
-            switch (l.type) {
+            switch (b.type) {
                 case BoudnaryType.ReturnStatement:
                     ++i
                     if (result.summary)
                         result.summary.end = i;
                     break main;
                 case BoudnaryType.Diagnostic: {
-                    const code = l.capture.substring(0, 8);
-                    const text = l.capture.substring(9).trim();
+                    const code = b.capture.substring(0, 8);
+                    const text = b.capture.substring(9).trim();
                     const d = new vscode.Diagnostic(line.range, text, asLevel(code));
                     d.code = code;
                     result.diagnostics.push(d);
                     break;
                 }
                 case BoudnaryType.OptionsRef:
-                    lastSection = updateCommonSection('options', i);
+                    lastListingSection = updateCommonSection('options', i);
                     break;
                 case BoudnaryType.ObjectCodeHeader: {
-                    if (l.capture.length < 45)
+                    if (b.capture.length < 45)
                         result.type = 'short';
                     else
                         result.type = 'long';
@@ -219,47 +291,50 @@ function processListing(doc: vscode.TextDocument, start: number, hasPrefix: bool
                         end: i + 1,
                         title: lastTitle,
                         codeStart: i + 1,
+                        dsect: b.capture.startsWith('D'),
+                        firstStmtNo: undefined,
                     };
-                    lastSection = codeSection;
+                    lastListingSection = codeSection;
                     result.codeSections.push(codeSection);
                     break;
                 }
                 case BoudnaryType.ExternalRef:
-                    lastSection = updateCommonSection('externals', i);
+                    lastListingSection = updateCommonSection('externals', i);
+                    state = States.ExternalRefs;
                     break;
                 case BoudnaryType.OrdinaryRef:
-                    lastSection = updateCommonSection('ordinary', i);
+                    lastListingSection = updateCommonSection('ordinary', i);
                     state = States.OrdinaryRefs;
                     ++i; // skip header
                     break;
                 case BoudnaryType.MacroRef:
-                    lastSection = updateCommonSection('macro', i);
+                    lastListingSection = updateCommonSection('macro', i);
                     state = States.Refs;
                     break;
                 case BoudnaryType.DsectRef:
-                    lastSection = updateCommonSection('dsects', i);
-                    state = States.Refs;
+                    lastListingSection = updateCommonSection('dsects', i);
+                    state = States.DsectRefs;
                     break;
                 case BoudnaryType.RegistersRef:
-                    lastSection = updateCommonSection('registers', i);
+                    lastListingSection = updateCommonSection('registers', i);
                     state = States.Refs;
                     break;
                 case BoudnaryType.DiagnosticRef:
-                    lastSection = updateCommonSection('summary', i);
+                    lastListingSection = updateCommonSection('summary', i);
                     state = States.Refs;
                     break;
                 case BoudnaryType.RelocationDict:
-                    lastSection = updateCommonSection('relocations', i);
+                    lastListingSection = updateCommonSection('relocations', i);
                     state = States.Refs;
                     break;
                 case BoudnaryType.UsingsRef:
-                    lastSection = updateCommonSection('usings', i);
-                    state = States.Refs;
+                    lastListingSection = updateCommonSection('usings', i);
+                    state = States.UsingRef;
                     break;
                 case BoudnaryType.OtherBoundary:
-                    if (state === States.Options)
+                    if (state === States.Options || state === States.ExternalRefs)
                         state = States.Code;
-                    lastTitle = l.capture.substring(9).trim();
+                    lastTitle = b.capture.substring(9).trim();
                     lastTitleLine = i;
                     break;
             }
@@ -267,7 +342,21 @@ function processListing(doc: vscode.TextDocument, start: number, hasPrefix: bool
         else if (state === States.Code) {
             const obj = (result.type === 'short' ? r.objShortCode : r.objLongCode).exec(line.text);
             if (obj) {
-                result.statementLines.set(parseInt(obj[1]), i);
+                const address = obj[1] !== undefined ? parseInt(obj[1], 16) : undefined;
+                const stmtNumber = parseInt(obj[3]);
+                const symbolCandidate = obj[4];
+                if (symbolCandidate) {
+                    symbolCandidates.push({ name: symbolCandidate.toUpperCase(), stmtNumber });
+                }
+                result.statementLines.set(stmtNumber, { listingLine: i, address });
+                if (stmtNumber > result.maxStmtNum)
+                    result.maxStmtNum = stmtNumber;
+                if (lastListingSection && 'codeStart' in lastListingSection && lastListingSection.codeStart === i)
+                    lastListingSection.firstStmtNo = stmtNumber;
+
+                if (address !== undefined && (result.type === 'short' ? r.sectionAddrShort : r.sectionAddrLong).test(obj[2])) {
+                    addrTransionsStmtNumbers.push(stmtNumber);
+                }
             }
         }
         else if (state === States.OrdinaryRefs) {
@@ -277,21 +366,34 @@ function processListing(doc: vscode.TextDocument, start: number, hasPrefix: bool
                 if (symbol)
                     updateSymbols(result.symbols, symbol);
 
-                symbol = new Symbol();
-                if (ref[1]) {
-                    symbol.name = ref[1];
-                    symbol.defined.push(+ref[4]);
-                    refs = ref[5];
-                }
-                else {
-                    symbol.name = ref[6];
-                }
+                symbol = {
+                    name: ref[1] ? ref[1] : ref[9],
+                    defined: ref[1] ? [+ref[7]] : [],
+                    references: [],
+                    referencesPure: [],
+                    details: ref[1] ? {
+                        value: parseInt(ref[3], 16),
+                        sectionId: parseInt(ref[4], 16) | 0,
+                        reloc: ref[5] === ' ',
+                        undefined: ref[6] === 'U', // EQU mostly
+                        loctr: ref[6] === 'J',
+                    } : undefined,
+                };
+
+                refs = ref[8];
             }
             else if (symbol) {
                 const alt = r.ordinaryRefAltSecondLine.exec(line.text);
                 if (alt) {
-                    symbol.defined.push(+alt[4]);
-                    refs = alt[5];
+                    symbol.details = {
+                        value: parseInt(alt[3], 16),
+                        sectionId: parseInt(alt[4], 16) | 0,
+                        reloc: alt[5] === ' ',
+                        undefined: alt[6] === 'U', // EQU mostly
+                        loctr: alt[6] === 'J',
+                    };
+                    symbol.defined.push(+alt[7]);
+                    refs = alt[8];
                 }
                 else {
                     const cont = r.ordinaryRefRest.exec(line.text);
@@ -302,17 +404,102 @@ function processListing(doc: vscode.TextDocument, start: number, hasPrefix: bool
             }
 
             if (refs && symbol) {
-                for (const m of refs.matchAll(/(\d+)[BDMUX]?/g)) {
-                    symbol.references.push(+m[1]);
+                for (const m of refs.matchAll(/(\d+)([BDMUX])?/g)) {
+                    const line = +m[1];
+                    symbol.references.push(line);
+                    if (!m[2])
+                        symbol.referencesPure.push(line);
                 }
+            }
+        }
+        else if (state === States.ExternalRefs) {
+            const ref = r.externalRefFirstLine.exec(line.text);
+            let data;
+            if (ref) {
+                lastCsect = ref[1];
+                if (ref[2]) {
+                    data = ref;
+                }
+            } else if (lastCsect) {
+                const rest = r.externalRefSecondLine.exec(line.text);
+                if (rest) {
+                    data = rest;
+                }
+            }
+            if (lastCsect && data && isCsectLike(data[2])) {
+                const ID = data[3];
+                const ownerName = data[6].startsWith(' ') ? lastCsect : (result.sections.get(data[6]) || lastCsect);
+                result.sections.set(ID, ownerName);
+                if (!data[4].startsWith(' ') && !data[5].startsWith(' ')) {
+                    const address = parseInt(data[4], 16);
+                    const length = parseInt(data[5], 16);
+                    result.csects.push({ name: ownerName, address, length });
+                }
+            }
+        }
+        else if (state === States.DsectRefs) {
+            const ref = r.dsectRefFirstLine.exec(line.text);
+            let data;
+            if (ref) {
+                lastDsect = ref[1] ? ref[1] : ref[5];
+                lastDsect = lastDsect.toUpperCase();
+                if (ref[2]) {
+                    data = ref;
+                }
+            } else if (lastDsect) {
+                const rest = r.dsectRefSecondLine.exec(line.text);
+                if (rest) {
+                    data = rest;
+                }
+            }
+            if (lastDsect && data && result.symbols.get(lastDsect) === undefined) {
+                symbol = {
+                    name: lastDsect,
+                    defined: [+data[4]],
+                    references: [],
+                    referencesPure: [],
+                    details: {
+                        value: 0,
+                        sectionId: parseInt(data[3], 16) | 0,
+                        reloc: true,
+                        undefined: false,
+                        loctr: true,
+                    },
+                };
+                result.symbols.set(lastDsect, symbol);
+            }
+        }
+        else if (state === States.UsingRef) {
+            const ref = r.usingMapLine.exec(line.text);
+            if (ref) {
+                const stmtNo = +ref[1];
+                const sectionId = parseInt(ref[3], 16) | 0;
+                if (sectionId >= 0)
+                    knownCsectStmtNumbers.push(stmtNo);
             }
         }
     }
     if (symbol)
         updateSymbols(result.symbols, symbol);
-    if (lastSection) {
-        lastSection.end = i;
+    if (lastListingSection) {
+        lastListingSection.end = i;
     }
+
+    for (const c of symbolCandidates) {
+        if (result.symbols.has(c.name)) continue;
+        result.symbols.set(c.name, {
+            name: c.name,
+            defined: [c.stmtNumber],
+            references: [],
+            referencesPure: [],
+            details: undefined,
+        });
+    }
+
+    result.sectionMap = createSectionMap(result, knownCsectStmtNumbers, addrTransionsStmtNumbers);
+
+    for (const s of result.symbols.values())
+        s.defined.sort((l, r) => l - r);
 
     result.end = i;
     return { nexti: i, result };
@@ -421,14 +608,134 @@ function codeSectionAsSymbol(s: Section & { title: string }) {
     return new vscode.DocumentSymbol(s.title ? s.title : '(untitled)', '', vscode.SymbolKind.Package, r, r);
 }
 
-function listingAsSymbol(l: Listing, id: number | undefined) {
-    const result = new vscode.DocumentSymbol(
+function labelAsSymbol(s: Symbol, pos: number) {
+    const r = new vscode.Range(pos, 0, pos, 0);
+    return new vscode.DocumentSymbol(s.name, '', vscode.SymbolKind.Object, r, r);
+}
+
+function listVisibleSymbolsOrdered(l: Listing) {
+    const visibleSymbols = [];
+    for (const s of l.symbols.values()) {
+        if (s.name.startsWith('=')) continue; // skip literals
+        for (const stmtNo of s.defined) {
+            const fileLine = l.statementLines.get(stmtNo)?.listingLine;
+            if (fileLine !== undefined) {
+                visibleSymbols.push({ symbol: s, line: fileLine });
+                break;
+            }
+        }
+    }
+    visibleSymbols.sort((l, r) => l.line - r.line);
+    return visibleSymbols;
+}
+
+function createSectionMap(l: Listing, knownCsectStmtNumbers: number[], addrTransitions: number[]) {
+    const limit = l.maxStmtNum;
+    const result: (boolean | undefined)[] = new Array(limit).fill(undefined);
+    const dsectRef = 1;
+    const csectRef = 2;
+    const loctrPureRef = 4;
+    const refCount = 8;
+    const hints = new Array(limit).fill(0);
+
+    for (const s of l.symbols.values()) {
+        const details = s.details;
+        if (!details) continue;
+        if (details.loctr) {
+            for (const r of s.referencesPure) {
+                if (r >= limit) continue;
+                hints[r] += refCount;
+                hints[r] |= details.sectionId < 0 ? dsectRef : csectRef | loctrPureRef;
+            }
+        }
+        if (details.undefined || !details.reloc) continue;
+        for (const d of s.defined) {
+            if (d >= limit) continue;
+            result[d] = details.sectionId >= 0;
+            hints[d] += refCount;
+            hints[d] |= details.sectionId < 0 ? dsectRef : csectRef;
+        }
+    }
+
+    for (const s of l.codeSections) {
+        const stmtNo = s.firstStmtNo;
+        if (stmtNo === undefined || stmtNo >= limit) continue;
+        result[stmtNo] = s.dsect == false;
+    }
+
+    for (const stmtNo of knownCsectStmtNumbers) {
+        if (stmtNo >= limit) continue;
+        result[stmtNo] = true;
+    }
+
+    const csectTransition = refCount | csectRef;
+    const dsectTransition = refCount | dsectRef;
+    for (const stmtNo of addrTransitions) {
+        if (stmtNo >= limit) continue;
+        const ignoreLoctr = hints[stmtNo] & ~loctrPureRef;
+        if (ignoreLoctr == csectTransition)
+            result[stmtNo] = true;
+        else if (ignoreLoctr == dsectTransition)
+            result[stmtNo] = false;
+    }
+
+    const lastValid = result.findIndex(e => e !== undefined);
+    if (lastValid >= 0) {
+        for (let i = lastValid + 1; i < result.length; ++i) {
+            if (result[i] !== undefined) continue;
+            result[i] = result[i - 1] || (hints[i] & loctrPureRef) != 0;
+        }
+    }
+
+    return result;
+}
+
+function provideObjectCodeDetails(l: Listing, code: vscode.DocumentSymbol) {
+    code.children = l.codeSections.reduce<typeof l.codeSections>((acc, c) => {
+        const last = acc[acc.length - 1];
+        if (last?.title === c.title) {
+            last.start = Math.min(last.start, c.start);
+            last.end = Math.max(last.end, c.end);
+        }
+        else {
+            acc.push(c);
+        }
+        return acc;
+    }, []).map(x => codeSectionAsSymbol(x));
+
+    const visibleSymbols = listVisibleSymbolsOrdered(l);
+
+    let titleId = -1;
+    let parent = code.children;
+    let titleIdLimit = code.children.length === 0 ? Number.MAX_SAFE_INTEGER : code.children[0].range.start.line;
+
+    for (const { symbol, line } of visibleSymbols) {
+
+        while (line >= titleIdLimit) {
+            ++titleId
+            parent = code.children[titleId].children;
+            if (titleId >= code.children.length - 1) {
+                titleIdLimit = Number.MAX_SAFE_INTEGER;
+                break;
+            }
+            titleIdLimit = code.children[titleId + 1].range.start.line;
+        }
+        parent.push(labelAsSymbol(symbol, line));
+    }
+}
+
+function makeListingRoot(l: Listing, id: number | undefined) {
+    return new vscode.DocumentSymbol(
         id ? `Listing ${id}` : 'Listing',
         '',
         vscode.SymbolKind.Module,
         new vscode.Range(l.start, 0, l.end, 0),
         new vscode.Range(l.start, 0, l.end, 0)
     );
+}
+
+function listingAsSymbol(l: Listing, id: number | undefined) {
+    const result = makeListingRoot(l, id);
 
     if (l.options) {
         result.children.push(sectionAsSymbol(l.options, 'High Level Assembler Option Summary'));
@@ -439,17 +746,7 @@ function listingAsSymbol(l: Listing, id: number | undefined) {
     if (l.codeSections.length > 0) {
         const code = sectionAsSymbol({ start: l.codeSections[0].start, end: l.codeSections[l.codeSections.length - 1].end }, 'Object Code');
         result.children.push(code);
-        code.children = l.codeSections.reduce<typeof l.codeSections>((acc, c) => {
-            const last = acc[acc.length - 1];
-            if (last?.title === c.title) {
-                last.start = Math.min(last.start, c.start);
-                last.end = Math.max(last.end, c.end);
-            }
-            else {
-                acc.push(c);
-            }
-            return acc;
-        }, []).map(x => codeSectionAsSymbol(x));
+        provideObjectCodeDetails(l, code);
     }
     if (l.relocations) {
         result.children.push(sectionAsSymbol(l.relocations, 'Relocation Dictionary'));
@@ -476,7 +773,34 @@ function listingAsSymbol(l: Listing, id: number | undefined) {
     return result;
 }
 
-function findBestFitLine(m: Map<number, number>, s: number): number | undefined {
+function listingAsOffsets(l: Listing, id: number | undefined) {
+    const result = makeListingRoot(l, id);
+    const offsets = result.children;
+
+    for (const [stmtNo, ll] of l.statementLines.entries()) {
+        const address = ll.address;
+        if (!l.sectionMap[stmtNo] || address === undefined) continue;
+
+        const csect = l.csects.find(x => x.address <= address && address < x.address + x.length);
+
+        let suffix = ''
+        if (csect)
+            suffix = ` (${csect.name}+${(address - csect.address).toString(16).toUpperCase().padStart(8, '0')})`;
+
+        const rng = new vscode.Range(ll.listingLine, 0, ll.listingLine, 0);
+        offsets.push(new vscode.DocumentSymbol(
+            `Offset ${address.toString(16).toUpperCase().padStart(8, '0')}${suffix}`,
+            '',
+            vscode.SymbolKind.Object,
+            rng,
+            rng,
+        ));
+    }
+
+    return result;
+}
+
+function findBestFitLine(m: Map<number, ListingLine>, s: number): number | undefined {
     let result;
 
     for (const k of m.keys()) {
@@ -485,7 +809,7 @@ function findBestFitLine(m: Map<number, number>, s: number): number | undefined 
     }
 
     if (result)
-        return m.get(result);
+        return m.get(result)?.listingLine;
     else
         return undefined;
 }
@@ -530,19 +854,19 @@ export function createListingServices(diagCollection?: vscode.DiagnosticCollecti
         handleListingContent,
         releaseListingContent,
         provideDefinition: symbolFunction((symbol, l, document) => symbol.defined
-            .map(x => l.statementLines.get(x) || findBestFitLine(l.statementLines, x))
+            .map(x => l.statementLines.get(x)?.listingLine || findBestFitLine(l.statementLines, x))
             .filter((x): x is number => typeof x === 'number')
             .sort(compareNumbers)
             .map(x => new vscode.Location(document.uri, getCodeRange(l, x))))
         ,
-        provideReferences: symbolFunction((symbol, l, document, context: vscode.ReferenceContext) => symbol.computeReferences(context.includeDeclaration)
-            .map(x => l.statementLines.get(x))
+        provideReferences: symbolFunction((symbol, l, document, context: vscode.ReferenceContext) => computeReferences(symbol, context.includeDeclaration)
+            .map(x => l.statementLines.get(x)?.listingLine)
             .filter((x): x is number => typeof x === 'number')
             .sort(compareNumbers)
             .map(x => new vscode.Location(document.uri, getCodeRange(l, x)))
         ),
         provideHover: symbolFunction((symbol, l, document) => asHover(symbol.defined
-            .map(x => l.statementLines.get(x))
+            .map(x => l.statementLines.get(x)?.listingLine)
             .filter((x): x is number => typeof x === 'number')
             .sort(compareNumbers)
             .map(x => document.lineAt(x).text)
@@ -550,6 +874,9 @@ export function createListingServices(diagCollection?: vscode.DiagnosticCollecti
         ),
         provideDocumentSymbols: (document: vscode.TextDocument) =>
             (listings.get(document.uri.toString()) ?? handleListingContent(document))?.map((l, id, ar) => listingAsSymbol(l, ar.length > 1 ? id + 1 : undefined))
+        ,
+        provideOffsets: (document: vscode.TextDocument) =>
+            (listings.get(document.uri.toString()) ?? handleListingContent(document))?.map((l, id, ar) => listingAsOffsets(l, ar.length > 1 ? id + 1 : undefined))
         ,
     };
 }
@@ -569,5 +896,11 @@ export function registerListingServices(context: vscode.ExtensionContext) {
     context.subscriptions.push(vscode.languages.registerDefinitionProvider(languageIdHlasmListing, services));
     context.subscriptions.push(vscode.languages.registerReferenceProvider(languageIdHlasmListing, services));
     context.subscriptions.push(vscode.languages.registerHoverProvider(languageIdHlasmListing, services));
+    // VSCode ordering - last one at the top
+    context.subscriptions.push(vscode.languages.registerDocumentSymbolProvider(
+        languageIdHlasmListing,
+        { provideDocumentSymbols: (document) => services.provideOffsets(document) },
+        { label: "Offsets" })
+    );
     context.subscriptions.push(vscode.languages.registerDocumentSymbolProvider(languageIdHlasmListing, services));
 }